Factor WCS into own service
