+++ django_services/cache/urls.py
@@ -5,5 +5,4 @@ from . import views
 urlpatterns = [
     url(r'^$', views.index, name="index"),
     url(r'^cache/(.*)$', views.get_resource, name="cache"),
-    url(r'^wcs', views.wcs, name="wcs")
 ]
