+++ django_services/django_services/urls.py
@@ -19,4 +19,5 @@ from django.contrib import admin
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
     url(r'^cache/', include('cache.urls')),
+    url(r'^wcs/', include('wcs.urls')),
 ]
