+++ django_services/wcs/models.py
@@ -1,0 +1,5 @@
+from __future__ import unicode_literals
+
+from django.db import models
+
+# Create your models here.
