+++ django_services/wcs/urls.py
@@ -1,0 +1,7 @@
+from django.conf.urls import url
+
+from . import views
+
+urlpatterns = [
+    url(r'^wcs', views.wcs, name="wcs")
+]
